Added support for PFlags.
Sigar is the method through which ProActive users monitor their
RMNodes (cpu, memory, of the underlying host).
PFlags is another measure that allows application running on the
host to expose some monitoring information (like PID, URL, etc.)
by just writing a file with <name> that will be exposed in a
PFlags MBean attribute as a map, containing <name>:<content> of
the file (created somehow by the process).

Signed-off-by: Mauricio Jost <mauricio.jost@activeeon.com>